requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
databases==0.8.0
dataclasses-json==0.6.3
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.5.0
easydict==1.11
email-validator==2.1.0.post1
fastai==2.7.13
fastapi==0.109.0
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.13.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
greenlet==3.0.3
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.17
langchain-openai==0.0.5
langcodes==3.3.0
langsmith==0.0.85
llama-hub==0.0.77
llama-index==0.9.40
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
openai==1.10.0
opencv-python==4.9.0.80
orjson==3.9.12
packaging==23.2
pandas==2.2.0
pillow==10.2.0
preshed==3.0.9
protobuf==4.25.2
psutil==5.9.8
psycopg==3.1.17
psycopg-binary==3.1.17
pyaml==23.12.0
pydantic==2.6.0
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
pytz==2023.4
PyYAML==6.0.1
redis==5.0.1
regex==2023.12.25
requests==2.31.0
retrying==1.3.4
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.51
srsly==2.4.8
starlette==0.35.1
sympy==1.12
tenacity==8.2.3
tensorboardX==2.6.2.2
thinc==8.2.2
threadpoolctl==3.2.0
tiktoken==0.5.2
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
wasabi==1.1.2
watchfiles==0.21.0
weasel==0.3.4
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
